refactor: decompose AttachedDisksModel into MountTableParser + DiskEventHandler
The 855-line @MainActor model had three concerns tangled together:

  1. /sbin/mount subprocess + statvfs + parsing (~150 lines)
  2. Structured-event → state transforms (~120 lines)
  3. Orchestration / @Published store + lifecycle

Concerns (1) and (2) are now namespace-style enums in
DiskJockeyApplication/Models/:

  - enum MountTableParser
      .enumerate(fsTypesOfInterest:) -> [AttachedDisk]
      .parseMountLine(_:fsTypesOfInterest:) -> AttachedDisk?
      .bsdName(from:) -> String
      .isWholeDiskBSD(_:) -> Bool
      .statvfsInfo(mountPath:fsType:volumeName:) -> [String: String]
      .forceUnmountStale(mountPath:bsd:)
      (+ private splitMountLine for the parse internals)

  - enum DiskEventHandler
      .fsTypeFromEventKind(_:) -> String?
      .applyEventInPlace(kind:fields:to:)
      .applyVolumeInfo(fields:to:)
      .materializeTotalSize(from:) -> UInt64?
      .fsckStatus(kind:fields:disk:) -> FsckStatus?

case-less enum so they can't be instantiated and no state can leak in;
matches the spirit of the original `private static func` methods on
the model.

AttachedDisksModel.refresh() and applyExtensionEvent now delegate to
those namespaces with one-line calls. The model's public surface is
unchanged — public methods (start / stop / refresh / removeDisk / forget
/ setStatus / applyExtensionEvent / applyLogLine) keep identical
signatures and behaviour.

The one test-side rename:

  AttachedDisksModel.parseMountLine(...) → MountTableParser.parseMountLine(...)

(was a `nonisolated static` reachable from tests; now lives in the
parser namespace alongside its siblings).

New test coverage:

  DiskJockeyTests/MountTableParserTests.swift — 11 cases:
    parseMountLine: RW / RO / legacy-RO / fstype-not-in-interest /
                    malformed / NTFS read-only.
    bsdName: dev-prefixed / bare / empty.
    isWholeDiskBSD: bare disks accepted, slices rejected, malformed
                    rejected including the `"disk"` edge case the
                    anchored regex correctly rejects.

  DiskJockeyTests/DiskEventHandlerTests.swift — 17 cases:
    fsTypeFromEventKind: known prefixes / unrelated kinds / empty.
    fsckStatus: volume.clean / volume.dirty / fsck.start /
                fsck.progress field decoding / fsck.done sets
                lastRepairedCount + lastAnomaliesFound /
                fsck.failed carries error / unknown kind nil.
    applyVolumeInfo: fsType + volume_name adoption / volume_name
                     placeholder rule (don't clobber a mount-table
                     name) / total_size materialisation /
                     stableIdentity from UUID / stableIdentity from
                     NTFS serial.
    materializeTotalSize: ext4 multiply / ntfs passthrough /
                          unknown fs / ext4 overflow → nil.
    applyEventInPlace dispatch: volume.info → volume handler /
                                fsck.progress → fsckStatus /
                                unknown kind no-op.

Net effect:
  AttachedDisksModel.swift: 855 → 562 lines (-293)
  MountTableParser.swift:   new, 183 lines
  DiskEventHandler.swift:   new, 153 lines
  Tests:                    +325 lines across two new files

Behaviour preserved across the move (verified by 139 tests across the
app + library suites still passing). Build succeeded.
