+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The README carries an abbreviated tail of this file (last ten dated sections).
 
 ## 2026-06-02
 
+- **`AttachedDisksModel` decomposed into `MountTableParser` + `DiskEventHandler`.** The 855-line `@MainActor` model had three concerns tangled together: `/sbin/mount` subprocess + statvfs + parsing (~150 lines), structured-event → state transforms (~120 lines), and the orchestration / `@Published` store itself. The two helper concerns are now `enum MountTableParser` and `enum DiskEventHandler` in `DiskJockeyApplication/Models/` — namespace-style (case-less `enum`) so they can't be instantiated and no state can leak in. `AttachedDisksModel.refresh()` and `applyExtensionEvent` now delegate to those namespaces with one-line calls; the model's public API is unchanged. The existing `AttachedDisksModelTests` updates one call site (`AttachedDisksModel.parseMountLine` → `MountTableParser.parseMountLine`) — that the test still passes after the rename is the proof of behaviour preservation. New `DiskJockeyTests/MountTableParserTests.swift` (11 cases) covers parseMountLine for RW / RO / legacy-RO / fstype-filter / malformed-line / NTFS, bsdName extraction (both dev-prefixed and bare), and isWholeDiskBSD on bare disks / slices / malformed input including the `disk` edge case the original regex rejects. New `DiskJockeyTests/DiskEventHandlerTests.swift` (17 cases) covers fsTypeFromEventKind for known and unknown prefixes, fsckStatus dispatch across every kind (volume.clean / dirty / fsck.start / fsck.progress / fsck.done / fsck.failed / unknown), applyVolumeInfo (fs adoption, volume_name placeholder rule, volume_name clobber-protection, stableIdentity from UUID + from ntfs serial), materializeTotalSize (ext4 multiply / ntfs passthrough / unknown / overflow), and applyEventInPlace dispatch (volume.info → volume handler / fsck.progress → fsckStatus / unknown kind no-op). Full app + library suites still green (139 tests across both).
 - **`MountableFileSystem` protocol + `MountedResourceRegistry<R>` generic.** EXT4FileSystem and NTFSFileSystem each declared `static let mountedResources = OSAllocatedUnfairLock<[ObjectIdentifier: MountedResource]>(initialState: [:])` and reimplemented the same four ops against it: register on load, remove on unload, resolve-the-sole-entry for `startCheck`/`startFormat`, and find-by-bsdName for the RepairXPCService. Same shape, two byte-near-identical implementations. Now there's one `public final class MountedResourceRegistry<Resource: MountedResource>: @unchecked Sendable` in `DiskJockeyLibrary/MountableFileSystem.swift`, exposing `register(_:_:)` / `remove(_:)` / `resolveSingle()` / `first(where:)` / `count` — the four operations the callers actually do, each one line at the call site. Plus `protocol MountedResource` (codifies the shared `bsdName + opLock` fields the per-FS structs already had) and `protocol MountableFileSystem` (associates each FS class with its registry handle). Twelve raw `withLock { map in … }` blocks collapse to one-line API calls; both `EXT4FileSystem.MountedResource` / `NTFSFileSystem.MountedResource` adopt the library protocol with zero field changes; both classes carry an `extension … : MountableFileSystem {}` conformance. Eight new tests in `DiskJockeyLibraryTests/MountableFileSystemTests.swift` cover register / replace / remove / idempotent-remove / `resolveSingle` ambiguity refusal / `first(where:)` predicate / concurrent 100-key register-and-remove / a stand-in `MountableFileSystem` conformance that proves the associated-type machinery compiles independent of the FSKit-extension targets. Full library suite 41/41 green.
 - **Per-volume item cache extracted into reusable `FileIDCache<Item>`.** Both `EXT4Volume` and `NTFSVolume` carried a `[UInt64: Item]` map with the same get-or-create-or-replace shape (validate cached entry's path + parentID; replace if mismatch — the load-bearing guard against inode reuse + hard-link aliasing + rust-fs-ext4 mkdir-reused-fileID). EXT4 used `OSAllocatedUnfairLock`; NTFS used `NSLock`. Now there's one `public final class FileIDCache<Item: AnyObject>` in `DiskJockeyLibrary/FileIDCache.swift`, backed by `OSAllocatedUnfairLock` — NTFS upgrades from `NSLock` to the modern Sendable-correct primitive as a side effect. API is `getOrCreate(id:validate:create:)` + `remove(id:)` + `count`; the per-FS validation rule (path + parentInode for ext4, path + parentRecordNumber for NTFS) lives in the caller's closure so the cache stays oblivious to per-flavour invariants. Net change inside the volumes: `item(forID:)` / `item(forRecordNumber:)` bodies collapse from a manual `withLock { … }` block to a single `items.getOrCreate(…)` call; `reclaimItem` collapses to `items.remove(id:)`. Eight new tests in `DiskJockeyLibraryTests/FileIDCacheTests.swift` cover empty-cache create, validate-true hit, validate-false replace, distinct-keys independence, removal, no-op remove on absent id, and concurrent stress (200 / 500 parallel calls). Full library suite 33/33 green.
 - **`EXT4Item` / `NTFSItem` collapsed into generic `FileSystemItem<Tag>`.** The two per-FS classes (37 + 36 lines) were structurally identical — an identity value, a path, an optional parent identity — diverging only in the integer width of the identity (`UInt32` inode vs `UInt64` MFT record). Now there's one `open class FileSystemItem<Tag: FileSystemTag>: FSItem` in `DiskJockeyLibrary/FileSystemItem.swift`, with `EXT4Tag` / `NTFSTag` phantom markers pinning the identity width per filesystem. Call sites in `EXT4Volume` / `NTFSVolume` keep their familiar `inode` / `parentInode` / `fileRecordNumber` / `parentRecordNumber` spelling via tag-constrained extensions that forward to the generic storage — the volume code didn't change. Phantom-tag distinctness means a stray cast can no longer feed an ext4 inode into an NTFS code path; the compiler refuses it. Eleven new unit tests in `DiskJockeyLibraryTests/FileSystemItemTests.swift` cover init round-trip, legacy forwarders, ID widths, root (parent=nil) shape, and static type distinctness; full library suite 25/25 green.
