+++ DiskJockeyTests/AttachedDisksModelTests.swift
@@ -1,7 +1,8 @@
 //
-//  AttachedDisksModelTests.swift — covers parseMountLine() so we don't
-//  need to spawn /sbin/mount in tests. The parser produces the
-//  isWritable flag the sidebar / detail UI render.
+//  AttachedDisksModelTests.swift — minimal smoke tests for the
+//  public model surface. The bulk of parsing + event-handler
+//  coverage lives in `MountTableParserTests` / `DiskEventHandlerTests`
+//  since those helpers now live in their own namespace enums.
 //
 
 import Testing
@@ -13,7 +14,7 @@ struct AttachedDisksModelTests {
     @Test func testParseRO() throws {
         let line = "/dev/disk5s2 on /Volumes/rootfs (ext4, local, read-only, fskit)"
         let disk = try #require(
-            AttachedDisksModel.parseMountLine(line, fsTypesOfInterest: Self.interesting)
+            MountTableParser.parseMountLine(line, fsTypesOfInterest: Self.interesting)
         )
         #expect(disk.isWritable == false)
         #expect(disk.fsType == "ext4")
@@ -24,7 +25,7 @@ struct AttachedDisksModelTests {
     @Test func testParseRW() throws {
         let line = "/dev/disk5s2 on /Volumes/rootfs (ext4, local, fskit)"
         let disk = try #require(
-            AttachedDisksModel.parseMountLine(line, fsTypesOfInterest: Self.interesting)
+            MountTableParser.parseMountLine(line, fsTypesOfInterest: Self.interesting)
         )
         #expect(disk.isWritable == true)
         #expect(disk.fsType == "ext4")
@@ -33,7 +34,7 @@ struct AttachedDisksModelTests {
     @Test func testParseLegacyROToken() throws {
         let line = "/dev/disk5s2 on /Volumes/rootfs (ext4, ro, fskit)"
         let disk = try #require(
-            AttachedDisksModel.parseMountLine(line, fsTypesOfInterest: Self.interesting)
+            MountTableParser.parseMountLine(line, fsTypesOfInterest: Self.interesting)
         )
         #expect(disk.isWritable == false)
     }
